+++ src/hooks/ready/registerGameSettings.js
@@ -302,6 +302,24 @@ export default function () {
     default: true,
   });
 
+  game.settings.register("ddb-importer", "munching-policy-use-iconizer", {
+    name: "ddb-importer.munching-policy-use-iconizer.name",
+    hint: "ddb-importer.munching-policy-use-iconizer.hint",
+    scope: "player",
+    config: false,
+    type: Boolean,
+    default: true,
+  });
+
+  game.settings.register("ddb-importer", "munching-policy-use-ddb-icons", {
+    name: "ddb-importer.munching-policy-use-ddb-icons.name",
+    hint: "ddb-importer.munching-policy-use-ddb-icons.hint",
+    scope: "player",
+    config: false,
+    type: Boolean,
+    default: true,
+  });
+
   game.settings.register("ddb-importer", "munching-policy-download-monster-images", {
     name: "ddb-importer.munching-policy-download-monster-images.name",
     hint: "ddb-importer.munching-policy-download-monster-images.hint",